make conv2d and argmin/argmax work on gpu
